ADD 3 RECIPES
It was added 3 recipes for this project. They were: Caramel Pie, Butter Cookies and Chinese Noodles.

Before this commit there was no recipes, just the index structure.

Now there's the structure, the recipes and it's images.

With that, I am able to complete my first practice of The Odin Project.